+++ .gitignore
@@ -22,3 +22,4 @@ Thumbs.db
 
 ## Text Editors
 .vscode
+.idea/
