+++ manubot/tests/test_imports.py
@@ -1,10 +1,12 @@
 
 def test_imports():
     import manubot.cite
-    import manubot.cite.util
     import manubot.cite.arxiv
     import manubot.cite.doi
     import manubot.cite.pubmed
     import manubot.cite.url
     import manubot.process.manuscript
-    assert isinstance(manubot.cite.util.citeproc_retrievers, dict)
+
+def assert_instance_type():
+    from manubot.cite.citekey import citeproc_retrievers
+    assert isinstance(citeproc_retrievers, dict)
